+++ back-end/reference-counter-ref/src/main/java/ar/lamansys/refcounterref/domain/counterreference/CounterReferenceInfoBo.java
@@ -32,5 +32,5 @@ public class CounterReferenceInfoBo {
 
     private List<Integer> fileIds;
 
-	private Integer closureTypeId;
+	private Short closureTypeId;
 }
