+++ back-end/reference-counter-ref/src/main/java/ar/lamansys/refcounterref/infraestructure/input/rest/dto/counterreference/CounterReferenceDto.java
@@ -35,5 +35,5 @@ public class CounterReferenceDto implements Serializable {
     private List<Integer> fileIds = new ArrayList<>();
 
 	@NotNull
-	private Integer closureTypeId;
+	private Short closureTypeId;
 }
